Add Maven Central repository and re-order property settings
This commit adds Maven Central as a repository in the gitblit-bom build.gradle file to ensure that dependencies can be resolved. Additionally, it fixes the Selenium version in the build.gradle to 2.28.0, replacing the dynamic resolution from gitblit-bom properties, providing more stability in dependency management. Also, the classpath setting in federationClassList task is adjusted to avoid potential configuration issues.